% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{app_server}
\alias{app_server}
\title{qPCRplots Shiny Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{User inputs.}

\item{output}{Application outputs.}

\item{session}{Session status.}
}
\description{
This is the server side logic of a Shiny web application.
Define server logic required by the app.
}
