% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_qpcrplots}
\alias{launch_qpcrplots}
\title{Launch qPCRplots Web App}
\usage{
launch_qpcrplots()
}
\description{
Launch qPCRplots Shiny Web Application Inspired and adapted from
Dean Attali answer https://stackoverflow.com/a/49623819/8796710
}
