% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Delta Ct Boxplot}
\usage{
plot_boxplot(data, control_group = NULL, notch_value = FALSE,
  cutoff = "-7", log_transform = FALSE, point_size = 2,
  text_size = 16)
}
\arguments{
\item{data}{Tibble or data frame with delta Ct values and
sample groups info}

\item{control_group}{Sample group used to compare statistics.
This group is used as the first group to plot.}

\item{notch_value}{If TRUE, make a notched box plot,
instead of normal boxplot. Default = FALSE.}

\item{cutoff}{Detection threshold for qPCR experiment.}

\item{log_transform}{If TRUE,log10-transform the delta Ct values.
Default = FALSE.}

\item{point_size}{Value for the size of plot elements.}

\item{text_size}{Value for the size of plot textual elements.}
}
\description{
Plots boxplot and statistics from qPCR experiments.
}
