% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrplots.R
\docType{package}
\name{qPCRplots}
\alias{qPCRplots}
\alias{qPCRplots-package}
\title{\code{qPCRplots} package}
\description{
Shiny app intended in helping process and visualize output from quantitative polymerase chain reaction experiment
}
\details{
See the README on
\href{https://github.com/luciorq/qPCRplots#readme}{GitHub}
}
